Bolla Nurse ni Faqat Excel bemorlar royhati qoldi xolos

+++ src/main/java/com/example/repository/PatientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface PatientRepository extends CrudRepository<PatientEntity, Integer> {
+public interface PatientRepository extends CrudRepository<PatientEntity, Long> {
     List<PatientEntity> getByFullNameIgnoreCase(String name);
 
     Integer countByFloor(String floor);
